// src/ohp-frontoffice/libs/mfe-broker/mfe-broker-portal/core/src/lib/models/fe-address.ts
export interface FeAddress {
  lineOne?: string | null | undefined;
  lineTwo?: string | null | undefined;
  lineThree?: string | null | undefined;
  lineFour?: string | null | undefined;
  lineFive?: string | null | undefined;
  postcode?: string | null | undefined;
  city?: string | null | undefined;
  country?: string | null | undefined;
}